+++ packages/Ecotone/tests/Projecting/ProjectingTest.php
@@ -7,25 +7,25 @@ declare(strict_types=1);
 
 namespace Test\Ecotone\Projecting;
 
+use Ecotone\EventSourcing\Attribute\ProjectionInitialization;
 use Ecotone\Lite\EcotoneLite;
 use Ecotone\Messaging\Attribute\Asynchronous;
 use Ecotone\Messaging\Channel\SimpleMessageChannelBuilder;
 use Ecotone\Messaging\Config\ModulePackageList;
 use Ecotone\Messaging\Config\ServiceConfiguration;
 use Ecotone\Messaging\Endpoint\ExecutionPollingMetadata;
+use Ecotone\Messaging\MessageHeaders;
 use Ecotone\Modelling\Attribute\EventHandler;
 use Ecotone\Modelling\Event;
 use Ecotone\Projecting\Attribute\Projection;
 use Ecotone\Projecting\InMemory\InMemoryStreamSourceBuilder;
-
-use function get_class;
-
 use PHPUnit\Framework\TestCase;
+use RuntimeException;
 
 /**
  * @internal
  */
-class ProjectingTestCase extends TestCase
+class ProjectingTest extends TestCase
 {
     public function test_asynchronous_projection(): void
     {
@@ -39,7 +39,7 @@ class ProjectingTestCase extends TestCase
             }
         };
         $ecotone = EcotoneLite::bootstrapFlowTesting(
-            [get_class($projection)],
+            [$projection::class],
             [$projection],
             configuration: ServiceConfiguration::createWithDefaults()
                 ->withSkippedModulePackageNames(ModulePackageList::allPackagesExcept([ModulePackageList::ASYNCHRONOUS_PACKAGE]))
@@ -70,7 +70,7 @@ class ProjectingTestCase extends TestCase
             }
         };
         $ecotone = EcotoneLite::bootstrapFlowTesting(
-            [get_class($projection)],
+            [$projection::class],
             [$projection],
             configuration: ServiceConfiguration::createWithDefaults()
                 ->withSkippedModulePackageNames(ModulePackageList::allPackages())
@@ -102,7 +102,7 @@ class ProjectingTestCase extends TestCase
             }
         };
         $ecotone = EcotoneLite::bootstrapFlowTesting(
-            [get_class($projection)],
+            [$projection::class],
             [$projection],
             configuration: ServiceConfiguration::createWithDefaults()
                 ->withSkippedModulePackageNames(ModulePackageList::allPackagesExcept([ModulePackageList::ASYNCHRONOUS_PACKAGE]))
@@ -123,5 +123,49 @@ class ProjectingTestCase extends TestCase
         $this->assertCount(0, $projection->handledEvents);
         $ecotone->run('async', ExecutionPollingMetadata::createWithTestingSetup());
         $this->assertCount(2, $projection->handledEvents);
+    }
+
+    public function test_it_can_init_projection_lifecycle_state(): void
+    {
+        $projection = new #[Projection(self::NAME)] class {
+            public const NAME = 'projection_with_lifecycle';
+            public const TICKET_CREATED = 'ticket.created';
+            private bool $initialized = false;
+            public array $projectedEvents = [];
+
+            #[EventHandler(self::TICKET_CREATED)]
+            public function on(array $event): void
+            {
+                if (! $this->initialized) {
+                    throw new RuntimeException('Projection not initialized');
+                }
+                $this->projectedEvents[] = $event;
+            }
+
+            #[ProjectionInitialization]
+            public function init(): void
+            {
+                if ($this->initialized) {
+                    throw new RuntimeException('Projection already initialized');
+                }
+                $this->initialized = true;
+            }
+        };
+
+        $ecotone = EcotoneLite::bootstrapFlowTesting(
+            [$projection::class],
+            [$projection],
+            ServiceConfiguration::createWithDefaults()
+                ->addExtensionObject($streamSource = new InMemoryStreamSourceBuilder())
+        );
+
+        $streamSource->append(
+            Event::createWithType($projection::TICKET_CREATED, [], [MessageHeaders::EVENT_AGGREGATE_ID => 'ticket-1']),
+            Event::createWithType($projection::TICKET_CREATED, [], [MessageHeaders::EVENT_AGGREGATE_ID => 'ticket-4']),
+        );
+        self::assertEquals([], $projection->projectedEvents);
+
+        $ecotone->publishEventWithRoutingKey($projection::TICKET_CREATED, [MessageHeaders::EVENT_AGGREGATE_ID => 'ticket-1']);
+        self::assertCount(2, $projection->projectedEvents);
     }
 }
