+++ packages/PdoEventSourcing/tests/Projecting/App/Ordering/Command/CancelOrder.php
@@ -1,4 +1,5 @@
 <?php
+
 /*
  * licence Enterprise
  */
