--- packages/PdoEventSourcing/tests/Projecting/Fixture/TicketAsynchronousProjection.php
@@ -1,34 +1,0 @@
-<?php
-
-/*
- * licence Enterprise
- */
-declare(strict_types=1);
-
-namespace Test\Ecotone\EventSourcing\Projecting\Fixture;
-
-use Ecotone\Messaging\Attribute\Asynchronous;
-use Ecotone\Messaging\MessageHeaders;
-use Ecotone\Modelling\Attribute\EventHandler;
-use Ecotone\Projecting\Attribute\Projection;
-use Test\Ecotone\EventSourcing\Projecting\Fixture\Ticket\TicketCreated;
-
-#[Projection(self::NAME, MessageHeaders::EVENT_AGGREGATE_ID)]
-#[Asynchronous(self::ASYNC_CHANNEL)]
-class TicketAsynchronousProjection
-{
-    public const NAME = 'projection_with_async';
-    public const ASYNC_CHANNEL = 'async_projection';
-
-    private array $projectedEvents = [];
-    #[EventHandler]
-    public function on(TicketCreated $event): void
-    {
-        $this->projectedEvents[] = $event;
-    }
-
-    public function getProjectedEvents(): array
-    {
-        return $this->projectedEvents;
-    }
-}
