--- packages/PdoEventSourcing/tests/Projecting/ProjectingTest.php
@@ -1,56 +1,0 @@
-<?php
-
-/*
- * licence Enterprise
- */
-declare(strict_types=1);
-
-namespace Test\Ecotone\EventSourcing\Projecting;
-
-use Ecotone\Modelling\Event;
-use Ecotone\Projecting\InMemory\InMemoryProjectionStateStorage;
-use Ecotone\Projecting\InMemory\InMemoryProjector;
-use Ecotone\Projecting\InMemory\InMemoryStreamSource;
-use Ecotone\Projecting\NullPartitionProvider;
-use Ecotone\Projecting\ProjectingManager;
-use Ecotone\Projecting\ProjectorExecutor;
-use PHPUnit\Framework\TestCase;
-
-/**
- * @internal
- */
-class ProjectingTest extends TestCase
-{
-    protected function buildProjectingManager(ProjectorExecutor $projectorExecutor): ProjectingManager
-    {
-        $source = new InMemoryStreamSource();
-        $source->append(
-            Event::createWithType('an-event-type', ['id' => 1]),
-            Event::createWithType('an-event-type', ['id' => 2]),
-            Event::createWithType('an-event-type', ['id' => 3]),
-        );
-        $projectionStateStorage = new InMemoryProjectionStateStorage();
-
-        return new ProjectingManager(
-            $projectionStateStorage,
-            $projectorExecutor,
-            $source,
-            new NullPartitionProvider(),
-            'projection-name',
-        );
-    }
-
-    public function test_projecting(): void
-    {
-        $projector = new InMemoryProjector();
-        $projectingManager = $this->buildProjectingManager($projector);
-
-        self::assertCount(0, $projector);
-
-        $projectingManager->execute();
-        self::assertCount(3, $projector);
-
-        $projectingManager->execute();
-        self::assertCount(3, $projector);
-    }
-}
